+++ app/[locale]/template/customizations/[templateCustomizationId]/q/new/page.tsx
@@ -324,6 +324,9 @@ const CustomQuestionNew: React.FC = () => {
             backUrl={routePath('template.customize.question.create', { templateCustomizationId }, { section_id: sectionId, step: 1 })}
             successUrl={routePath('template.customize', { templateCustomizationId })}
             onSave={async (commonFields) => {
+              // Filter out displayOrder from the commonFields since it's not needed for the mutation input and is causing issues with the type.
+              // We will set displayOrder in the backend based on the pinnedQuestionId and pinnedQuestionType that we are passing in to pin the 
+              // new question right after the last question in the section
               const { displayOrder: _ignored, ...questionFields } = commonFields;
               const input = {
                 templateCustomizationId: Number(templateCustomizationId),